+++ http_stubs/admin.py
@@ -10,9 +10,10 @@ admin.site.site_title = 'Kesha'
 admin.site.site_header = 'Kesha Admin'
 
 
-@admin.register(models.LogEntry)
-class LogEntryAdmin(admin.ModelAdmin):
-    """Log entries admin."""
+class LogEntryAdminBase(admin.ModelAdmin):
+    """Base admin for LogEntry and ProxyLogEntry."""
+
+    change_form_template = 'admin/http_stubs/logentry/change_form.html'
 
     def has_add_permission(self, *args, **kwargs) -> bool:
         """Forbids adding new entries.
@@ -23,42 +24,113 @@ class LogEntryAdmin(admin.ModelAdmin):
         """
         return False
 
-    def pretty_body(self, instance) -> str:
-        """Jsonify the request body if possible.
+    def pretty_str(self, string: str) -> str:
+        """Jsonify a string if possible.
 
-        :param instance: instance of LogEntry
-        :returns: jsonify body
+        :param string: a string
+        :returns: jsonify string
         """
         try:
-            body = json.loads(instance.body)
+            body = json.loads(string)
         except json.JSONDecodeError:
-            body = None
+            body = ''
         else:
             body = json.dumps(body, indent=2)
         return body
 
-    pretty_body.short_description = 'Jsonify request body'
+    def pretty_request_body(self, instance) -> str:
+        """Jsonify the request body if possible.
 
-    list_filter = ('date', 'method')
+        :param instance: instance of a log entry
+        :returns: jsonify body
+        """
+        return self.pretty_str(instance.request_body)
+
+    pretty_request_body.short_description = 'Jsonify request body'
+
+    list_filter = ('request_date', 'method')
     search_fields = ('path', 'source_ip')
-    list_display = ('pk', 'date', 'http_stub', 'source_ip')
+    list_display = ('pk', 'request_date', 'http_stub', 'source_ip')
     readonly_fields = (
         'pk',
-        'date',
+        'request_date',
         'http_stub',
+        'resp_status',
+        'source_ip',
+        'result_script',
         'path',
         'method',
-        'source_ip',
-        'headers',
-        'body',
-        'pretty_body',
-        'result_script',
+        'request_headers',
+        'request_body',
+        'pretty_request_body',
+    )
+
+    fieldsets = (
+        (None, {
+            'fields': (
+                'pk',
+                'request_date',
+                'http_stub',
+                'resp_status',
+                'source_ip',
+                'result_script',
+            ),
+        }),
+        ('Request to Kesha', {
+            'fields': (
+                'path',
+                'method',
+                'request_headers',
+                'request_body',
+                'pretty_request_body',
+            ),
+        }),
     )
 
 
-@admin.register(models.HTTPStub)
-class HTTPStubAdmin(admin.ModelAdmin):
-    """HTTP stub admin."""
+@admin.register(models.LogEntry)
+class LogEntryAdmin(LogEntryAdminBase):
+    """Log entries admin."""
+
+
+@admin.register(models.ProxyLogEntity)
+class ProxyLogEntryAdmin(LogEntryAdminBase):
+    """Proxy log entries admin."""
+
+    def pretty_response_body(self, instance) -> str:
+        """Jsonify the response body if possible.
+
+        :param instance: instance of a log entry
+        :returns: jsonify body
+        """
+        return self.pretty_str(instance.response_body)
+
+    pretty_response_body.short_description = 'Jsonify response body'
+
+    readonly_fields = LogEntryAdminBase.readonly_fields + (
+        'target_path',
+        'response_latency',
+        'response_headers',
+        'response_body',
+        'pretty_response_body',
+    )
+
+    fieldsets = (
+        *LogEntryAdminBase.fieldsets,
+        ('Response from target endpoint', {
+            'fields': (
+                'target_path',
+                'response_latency',
+                'response_headers',
+                'response_body',
+                'pretty_response_body',
+            ),
+        }),
+    )
+
+
+class HTTPStubAdminBase(admin.ModelAdmin):
+    """Base admin for HTTPStub and ProxyHTTPStub."""
 
     change_form_template = 'admin/http_stubs/httpstub/change_form.html'
     extra_buttons_style = 'background-color:#00b0ff;color:white'
@@ -91,4 +163,40 @@ class HTTPStubAdmin(admin.ModelAdmin):
 
     list_display = ('pk', 'is_active', 'method', 'path')
 
-    list_filter = ('is_active', 'path', 'method', 'resp_status')
+    list_filter = ('is_active', 'path', 'method')
+
+
+@admin.register(models.HTTPStub)
+class HTTPStubAdmin(HTTPStubAdminBase):
+    """HTTP stub admin."""
+
+    list_filter = HTTPStubAdminBase.list_filter + ('resp_status',)
+
+
+@admin.register(models.ProxyHTTPStub)
+class ProxyHTTPStubAdmin(HTTPStubAdminBase):
+    """Proxy HTTP stub admin."""
+
+    fieldsets = (
+        (None, {
+            'fields': (
+                'is_active',
+                'path',
+                'regex_path',
+                'method',
+                'request_script',
+                'enable_logging',
+            ),
+        }),
+        ('Target settings', {
+            'fields': (
+                'target_url',
+                'allow_forward_query',
+                'target_ssl_verify',
+                'target_timeout',
+                'target_method',
+                'target_headers',
+                'target_body',
+            ),
+        }),
+    )