+++ http_stubs/tests/test_tasks.py
@@ -1,4 +1,7 @@
+from unittest.mock import patch
+
 import pytest
+from billiard.exceptions import SoftTimeLimitExceeded
 
 from http_stubs.tasks import run_request_script
 
@@ -7,13 +10,14 @@ from http_stubs.tasks import run_request_script
     'script, expect', (
         # valid scripts
         ('a = 1', 'Done'),
-        ("""
-def func(*args, **kwargs):
-    return json.dumps({'fake': 'data'})
-
-parsed_json = json.loads(func())
-assert parsed_json['fake'] == 'data'
-        """, 'Done'),  # noqa: WPS319
+        (
+            'def func(*args, **kwargs):\n'
+            "    return json.dumps({'fake': 'data'})\n"
+            '\n'
+            'parsed_json = json.loads(func())\n'
+            "assert parsed_json['fake'] == 'data'\n",
+            'Done',
+        ),
 
         # scripts with syntax error
         ('a = no_def', "Error: name 'no_def' is not defined"),
@@ -45,3 +49,17 @@ def test_run_without_log(script):
     :param script: a request script
     """
     run_request_script.run(script=script, request_body='')
+
+
+@patch('http_stubs.tasks.compile_restricted')
+def test_run_request_script_time_limit(compile_restricted, log_entity_factory):
+    """Test SoftTimeLimitExceeded Exception in run_request_script.
+
+    :param compile_restricted: patched compile_restricted func
+    :param log_entity_factory: factory log models
+    """
+    compile_restricted.side_effect = SoftTimeLimitExceeded
+    log = log_entity_factory()
+    run_request_script.delay(script='', request_body='', log_id=log.id)
+    log.refresh_from_db()
+    assert log.result_script == 'Error: Execution time limit exceeded'