+++ kesha/settings.py
@@ -9,7 +9,7 @@ environ.Env.read_env(str(BASE_DIR / '.env'))
 
 SECRET_KEY = env('KESHA_SECRET_KEY')
 
-DEBUG = env.bool('DEBUG', False)
+DEBUG = env.bool('KESHA_DEBUG', False)
 
 ALLOWED_HOSTS = env.list('KESHA_ALLOWED_HOSTS', default=['127.0.0.1'])
 
@@ -105,14 +105,18 @@ STATIC_ROOT = BASE_DIR / Path('static_build')
 # Admin UI settings
 SIMPLEUI_HOME_INFO = False
 SIMPLEUI_STATIC_OFFLINE = True
-SIMPLEUI_DEFAULT_THEME = 'layui.css'
+SIMPLEUI_DEFAULT_THEME = 'dark.green.css'
 SIMPLEUI_ANALYSIS = False
 SIMPLEUI_LOGO = STATIC_URL + 'kesha_icon.png'
 
+STUB_ICON = 'fas fa-feather-alt'
+
 SIMPLEUI_ICON = {
-    'HTTP Stubs': 'fas fa-feather-alt',
-    'Stubs': 'fas fa-feather-alt',
-    'Logs': 'fas fa-layer-group',
+    'HTTP Stubs': STUB_ICON,
+    'Request stubs': STUB_ICON,
+    'Request logs': 'fas fa-layer-group',
+    'Proxy stubs': STUB_ICON,
+    'Proxy logs': 'fas fa-layer-group',
 }
 
 INTERNAL_IPS = env.list('KESHA_INTERNAL_HOSTS', default=['127.0.0.1'])